PING_PATTERN was totally wrong.

+++ tests/handler_parser_tests.c
@@ -13,8 +13,8 @@ int test_parse(const char *test, int target_count)
 
     int rc = HandlerParser_execute(&parser, bdata(T1), blength(T1));
     debug("BODY %d long at %s with rc: %d, uuid: %s, target_count: %d",
-            parser.body_length, parser.body_start, rc, 
-            bdata(parser.uuid), parser.target_count);
+            (int)parser.body_length, parser.body_start, rc, 
+            bdata(parser.uuid), (int)parser.target_count);
 
     bdestroy(T1);
